remove logic and control methods from component to a separate class include channelId in interface
