+++ io.openems.edge.battery.soltaro/src/io/openems/edge/battery/soltaro/single/versionb/SingleRackVersionBImpl.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.osgi.service.cm.ConfigurationAdmin;
@@ -31,6 +30,7 @@ import io.openems.edge.battery.api.Battery;
 import io.openems.edge.battery.soltaro.ChannelIdImpl;
 import io.openems.edge.battery.soltaro.ModuleParameters;
 import io.openems.edge.battery.soltaro.single.versionb.statemachine.Context;
+import io.openems.edge.battery.soltaro.single.versionb.statemachine.ControlAndLogic;
 import io.openems.edge.battery.soltaro.single.versionb.statemachine.State;
 import io.openems.edge.bridge.modbus.api.AbstractOpenemsModbusComponent;
 import io.openems.edge.bridge.modbus.api.BridgeModbus;
@@ -48,8 +48,6 @@ import io.openems.edge.bridge.modbus.api.task.Task;
 import io.openems.edge.common.channel.Channel;
 import io.openems.edge.common.channel.IntegerDoc;
 import io.openems.edge.common.channel.IntegerReadChannel;
-import io.openems.edge.common.channel.IntegerWriteChannel;
-import io.openems.edge.common.channel.StateChannel;
 import io.openems.edge.common.component.OpenemsComponent;
 import io.openems.edge.common.event.EdgeEventConstants;
 import io.openems.edge.common.modbusslave.ModbusSlave;
@@ -69,11 +67,10 @@ import io.openems.edge.common.taskmanager.Priority;
 public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 		implements Battery, OpenemsComponent, EventHandler, ModbusSlave, StartStoppable, SingleRackVersionB {
 
-
 	private static final String KEY_TEMPERATURE = "_TEMPERATURE";
 	private static final String KEY_VOLTAGE = "_VOLTAGE";
 	private static final String NUMBER_FORMAT = "%03d"; // creates string number with leading zeros
-
+	
 	@Reference
 	protected ConfigurationAdmin cm;
 
@@ -85,8 +82,6 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 	private final StateMachine<State, Context> stateMachine = new StateMachine<>(State.UNDEFINED);
 	
 	private final Logger log = LoggerFactory.getLogger(SingleRackVersionBImpl.class);
-	private String modbusBridgeId;
-	
 	
 	private Config config;
 	private Map<String, Channel<?>> channelMap;
@@ -97,7 +92,7 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 				OpenemsComponent.ChannelId.values(), //
 				Battery.ChannelId.values(), //
 				StartStoppable.ChannelId.values(), //
-				SingleRackChannelId.values() //
+				SingleRackVersionB.ChannelId.values() //
 		);
 	}
 
@@ -116,21 +111,15 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 
 		super.activate(context, config.id(), config.alias(), config.enabled(), config.modbusUnitId(), this.cm, "Modbus",
 				config.modbus_id());
-		this.modbusBridgeId = config.modbus_id();
 
-		this.setWatchdog(config.watchdog());
-		this.setSoCLowAlarm(config.SoCLowAlarm());
-		this.setCapacity();
+		ControlAndLogic.setWatchdog(this, config.watchdog());
+		ControlAndLogic.setSoCLowAlarm(this, config.SoCLowAlarm());
+		ControlAndLogic.setCapacity(this, config);
 	}
-
-	private void setCapacity() {
-		int capacity = this.config.numberOfSlaves() * this.config.moduleType().getCapacity_Wh();
-		this._setCapacity(capacity);
-	}
-
+	
 	private void handleStateMachine() {
 		// Store the current State
-		this.channel(SingleRackChannelId.STATE_MACHINE).setNextValue(this.stateMachine.getCurrentState());
+		this.channel(SingleRackVersionB.ChannelId.STATE_MACHINE).setNextValue(this.stateMachine.getCurrentState());
 
 		// Initialize 'Start-Stop' Channel
 		this._setStartStop(StartStop.UNDEFINED);
@@ -142,15 +131,81 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 		try {
 			this.stateMachine.run(context);
 
-			this.channel(SingleRackChannelId.RUN_FAILED).setNextValue(false);
+			this.channel(SingleRackVersionB.ChannelId.RUN_FAILED).setNextValue(false);
 
 		} catch (OpenemsNamedException e) {
-			this.channel(SingleRackChannelId.RUN_FAILED).setNextValue(true);
+			this.channel(SingleRackVersionB.ChannelId.RUN_FAILED).setNextValue(true);
 			this.logError(this.log, "StateMachine failed: " + e.getMessage());
 		}
 	}
 
+	@Deactivate
+	protected void deactivate() {
+		// Remove dynamically created channels when component is deactivated
+		for (Channel<?> c : this.channelMap.values()) {
+			this.removeChannel(c);
+		}
+		super.deactivate();
+	}
 
+	@Override
+	public void handleEvent(Event event) {
+		if (!this.isEnabled()) {
+			return;
+		}
+		switch (event.getTopic()) {
+
+		case EdgeEventConstants.TOPIC_CYCLE_AFTER_PROCESS_IMAGE:
+			this.handleStateMachine();
+			break;
+		}
+	}
+	
+	@Override
+	public ModbusSlaveTable getModbusSlaveTable(AccessMode accessMode) {
+		return new ModbusSlaveTable(//
+				OpenemsComponent.getModbusSlaveNatureTable(accessMode), //
+				Battery.getModbusSlaveNatureTable(accessMode) //
+		);
+	}
+
+
+	@Override
+	public void setStartStop(StartStop value) {
+		if (this.startStopTarget.getAndSet(value) != value) {
+			// Set only if value changed
+			this.stateMachine.forceNextState(State.UNDEFINED);
+		}
+	}
+	
+	@Override
+	public StartStop getStartStopTarget() {
+		switch (this.config.startStop()) {
+		case AUTO:
+			// read StartStop-Channel
+			return this.startStopTarget.get();
+
+		case START:
+			// force START
+			return StartStop.START;
+
+		case STOP:
+			// force STOP
+			return StartStop.STOP;
+		}
+
+		assert false;
+		return StartStop.UNDEFINED; // can never happen
+	}
+
+	@Override
+	public String debugLog() {
+		return "SoC:" + this.getSoc() //
+				+ "|Discharge:" + this.getDischargeMinVoltage() + ";" + this.getDischargeMaxCurrent() //
+				+ "|Charge:" + this.getChargeMaxVoltage() + ";" + this.getChargeMaxCurrent() //
+				+ "|State:" + this.stateMachine.getCurrentState();
+	}
+	
 	/*
 	 * creates a map containing channels for voltage and temperature depending on
 	 * the number of modules
@@ -188,246 +243,114 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 		return "CLUSTER_1_BATTERY_" + String.format(NUMBER_FORMAT, num);
 	}
 
-	private void setWatchdog(int time_seconds) {
-		try {
-			IntegerWriteChannel c = this.channel(SingleRackChannelId.EMS_COMMUNICATION_TIMEOUT);
-			c.setNextWriteValue(time_seconds);
-		} catch (OpenemsNamedException e) {
-			this.log.error("Error while setting ems timeout!\n" + e.getMessage());
-		}
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		// Remove dynamically created channels when component is deactivated
-		for (Channel<?> c : this.channelMap.values()) {
-			this.removeChannel(c);
-		}
-		super.deactivate();
-	}
-
-	@Override
-	public void handleEvent(Event event) {
-		if (!this.isEnabled()) {
-			return;
-		}
-		switch (event.getTopic()) {
-
-		case EdgeEventConstants.TOPIC_CYCLE_AFTER_PROCESS_IMAGE:
-			this.handleStateMachine();
-			break;
-		}
-	}
-
-	private boolean isAlarmLevel2Error() {
-		return (this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CHA_CURRENT_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_LOW)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_LOW)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_DISCHA_CURRENT_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_CHA_TEMP_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_CHA_TEMP_LOW)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_SOC_LOW)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_TEMPERATURE_DIFFERENCE_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_POLES_TEMPERATURE_DIFFERENCE_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_DIFFERENCE_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_INSULATION_LOW)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_DIFFERENCE_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_DISCHA_TEMP_HIGH)
-				|| this.readValueFromBooleanChannel(SingleRackChannelId.ALARM_LEVEL_2_CELL_DISCHA_TEMP_LOW));
-	}
-
-	private boolean isSlaveCommunicationError() {
-		boolean b = false;
-		switch (this.config.numberOfSlaves()) {
-		case 20:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_20_COMMUNICATION_ERROR);
-		case 19:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_19_COMMUNICATION_ERROR);
-		case 18:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_18_COMMUNICATION_ERROR);
-		case 17:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_17_COMMUNICATION_ERROR);
-		case 16:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_16_COMMUNICATION_ERROR);
-		case 15:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_15_COMMUNICATION_ERROR);
-		case 14:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_14_COMMUNICATION_ERROR);
-		case 13:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_13_COMMUNICATION_ERROR);
-		case 12:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_12_COMMUNICATION_ERROR);
-		case 11:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_11_COMMUNICATION_ERROR);
-		case 10:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_10_COMMUNICATION_ERROR);
-		case 9:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_9_COMMUNICATION_ERROR);
-		case 8:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_8_COMMUNICATION_ERROR);
-		case 7:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_7_COMMUNICATION_ERROR);
-		case 6:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_6_COMMUNICATION_ERROR);
-		case 5:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_5_COMMUNICATION_ERROR);
-		case 4:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_4_COMMUNICATION_ERROR);
-		case 3:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_3_COMMUNICATION_ERROR);
-		case 2:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_2_COMMUNICATION_ERROR);
-		case 1:
-			b = b || this.readValueFromBooleanChannel(SingleRackChannelId.SLAVE_1_COMMUNICATION_ERROR);
-		}
-
-		return b;
-	}
-
-	public boolean hasError() {
-		return this.isAlarmLevel2Error() || this.isSlaveCommunicationError();
-	}
-
-	private boolean readValueFromBooleanChannel(SingleRackChannelId singleRackChannelId) {
-		StateChannel r = this.channel(singleRackChannelId);
-		Optional<Boolean> bOpt = r.value().asOptional();
-		return bOpt.isPresent() && bOpt.get();
-	}
-
-	public String getModbusBridgeId() {
-		return this.modbusBridgeId;
-	}
-
-	@Override
-	public String debugLog() {
-		return "SoC:" + this.getSoc() //
-				+ "|Discharge:" + this.getDischargeMinVoltage() + ";" + this.getDischargeMaxCurrent() //
-				+ "|Charge:" + this.getChargeMaxVoltage() + ";" + this.getChargeMaxCurrent() //
-				+ "|State:" + this.stateMachine.getCurrentState();
-	}
-
-
-	private void setSoCLowAlarm(int soCLowAlarm) {
-		try {
-			((IntegerWriteChannel) this.channel(SingleRackChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION))
-					.setNextWriteValue(soCLowAlarm);
-			((IntegerWriteChannel) this.channel(SingleRackChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION_RECOVER))
-					.setNextWriteValue(soCLowAlarm);
-		} catch (OpenemsNamedException e) {
-			this.log.error("Error while setting parameter for soc low protection!" + e.getMessage());
-		}
-	}
-
 	@Override
 	protected ModbusProtocol defineModbusProtocol() {
 
 		ModbusProtocol protocol = new ModbusProtocol(this, //
 				// Main switch
 				new FC6WriteRegisterTask(0x2010,
-						m(SingleRackChannelId.BMS_CONTACTOR_CONTROL, new UnsignedWordElement(0x2010)) //
+						m(SingleRackVersionB.ChannelId.BMS_CONTACTOR_CONTROL, new UnsignedWordElement(0x2010)) //
 				),
 
 				// System reset
 				new FC6WriteRegisterTask(0x2004, //
-						m(SingleRackChannelId.SYSTEM_RESET, new UnsignedWordElement(0x2004)) //
+						m(SingleRackVersionB.ChannelId.SYSTEM_RESET, new UnsignedWordElement(0x2004)) //
 				),
 
 				// EMS timeout --> Watchdog
 				new FC6WriteRegisterTask(0x201C,
-						m(SingleRackChannelId.EMS_COMMUNICATION_TIMEOUT, new UnsignedWordElement(0x201C)) //
+						m(SingleRackVersionB.ChannelId.EMS_COMMUNICATION_TIMEOUT, new UnsignedWordElement(0x201C)) //
 				),
 				// Sleep
-				new FC6WriteRegisterTask(0x201D, m(SingleRackChannelId.SLEEP, new UnsignedWordElement(0x201D)) //
+				new FC6WriteRegisterTask(0x201D, m(SingleRackVersionB.ChannelId.SLEEP, new UnsignedWordElement(0x201D)) //
 				),
 
 				// Work parameter
 				new FC6WriteRegisterTask(0x20C1,
-						m(SingleRackChannelId.WORK_PARAMETER_PCS_COMMUNICATION_RATE, new UnsignedWordElement(0x20C1)) //
+						m(SingleRackVersionB.ChannelId.WORK_PARAMETER_PCS_COMMUNICATION_RATE, new UnsignedWordElement(0x20C1)) //
 				),
 
 				// Paramaeters for configuring
 				new FC6WriteRegisterTask(0x2014,
-						m(SingleRackChannelId.AUTO_SET_SLAVES_ID, new UnsignedWordElement(0x2014))),
+						m(SingleRackVersionB.ChannelId.AUTO_SET_SLAVES_ID, new UnsignedWordElement(0x2014))),
 				new FC6WriteRegisterTask(0x2019,
-						m(SingleRackChannelId.AUTO_SET_SLAVES_TEMPERATURE_ID, new UnsignedWordElement(0x2019))),
+						m(SingleRackVersionB.ChannelId.AUTO_SET_SLAVES_TEMPERATURE_ID, new UnsignedWordElement(0x2019))),
 
 				// Control registers
 				new FC3ReadRegistersTask(0x2000, Priority.HIGH, //
-						m(SingleRackChannelId.FAN_STATUS, new UnsignedWordElement(0x2000)), //
-						m(SingleRackChannelId.MAIN_CONTACTOR_STATE, new UnsignedWordElement(0x2001)), //
-						m(SingleRackChannelId.DRY_CONTACT_1_EXPORT, new UnsignedWordElement(0x2002)), //
-						m(SingleRackChannelId.DRY_CONTACT_2_EXPORT, new UnsignedWordElement(0x2003)), //
-						m(SingleRackChannelId.SYSTEM_RESET, new UnsignedWordElement(0x2004)), //
-						m(SingleRackChannelId.SYSTEM_RUN_MODE, new UnsignedWordElement(0x2005)), //
-						m(SingleRackChannelId.PRE_CONTACTOR_STATUS, new UnsignedWordElement(0x2006)), //
+						m(SingleRackVersionB.ChannelId.FAN_STATUS, new UnsignedWordElement(0x2000)), //
+						m(SingleRackVersionB.ChannelId.MAIN_CONTACTOR_STATE, new UnsignedWordElement(0x2001)), //
+						m(SingleRackVersionB.ChannelId.DRY_CONTACT_1_EXPORT, new UnsignedWordElement(0x2002)), //
+						m(SingleRackVersionB.ChannelId.DRY_CONTACT_2_EXPORT, new UnsignedWordElement(0x2003)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_RESET, new UnsignedWordElement(0x2004)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_RUN_MODE, new UnsignedWordElement(0x2005)), //
+						m(SingleRackVersionB.ChannelId.PRE_CONTACTOR_STATUS, new UnsignedWordElement(0x2006)), //
 						m(new BitsWordElement(0x2007, this) //
-								.bit(15, SingleRackChannelId.ALARM_FLAG_STATUS_DISCHARGE_TEMPERATURE_LOW) //
-								.bit(14, SingleRackChannelId.ALARM_FLAG_STATUS_DISCHARGE_TEMPERATURE_HIGH) //
-								.bit(13, SingleRackChannelId.ALARM_FLAG_STATUS_VOLTAGE_DIFFERENCE) //
-								.bit(12, SingleRackChannelId.ALARM_FLAG_STATUS_INSULATION_LOW) //
-								.bit(11, SingleRackChannelId.ALARM_FLAG_STATUS_CELL_VOLTAGE_DIFFERENCE) //
-								.bit(10, SingleRackChannelId.ALARM_FLAG_STATUS_ELECTRODE_TEMPERATURE_HIGH) //
-								.bit(9, SingleRackChannelId.ALARM_FLAG_STATUS_TEMPERATURE_DIFFERENCE) //
-								.bit(8, SingleRackChannelId.ALARM_FLAG_STATUS_SOC_LOW) //
-								.bit(7, SingleRackChannelId.ALARM_FLAG_STATUS_CELL_OVER_TEMPERATURE) //
-								.bit(6, SingleRackChannelId.ALARM_FLAG_STATUS_CELL_LOW_TEMPERATURE) //
-								.bit(5, SingleRackChannelId.ALARM_FLAG_STATUS_DISCHARGE_OVER_CURRENT) //
-								.bit(4, SingleRackChannelId.ALARM_FLAG_STATUS_SYSTEM_LOW_VOLTAGE) //
-								.bit(3, SingleRackChannelId.ALARM_FLAG_STATUS_CELL_LOW_VOLTAGE) //
-								.bit(2, SingleRackChannelId.ALARM_FLAG_STATUS_CHARGE_OVER_CURRENT) //
-								.bit(1, SingleRackChannelId.ALARM_FLAG_STATUS_SYSTEM_OVER_VOLTAGE) //
-								.bit(0, SingleRackChannelId.ALARM_FLAG_STATUS_CELL_OVER_VOLTAGE) //
+								.bit(15, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_DISCHARGE_TEMPERATURE_LOW) //
+								.bit(14, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_DISCHARGE_TEMPERATURE_HIGH) //
+								.bit(13, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_VOLTAGE_DIFFERENCE) //
+								.bit(12, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_INSULATION_LOW) //
+								.bit(11, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_CELL_VOLTAGE_DIFFERENCE) //
+								.bit(10, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_ELECTRODE_TEMPERATURE_HIGH) //
+								.bit(9, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_TEMPERATURE_DIFFERENCE) //
+								.bit(8, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_SOC_LOW) //
+								.bit(7, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_CELL_OVER_TEMPERATURE) //
+								.bit(6, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_CELL_LOW_TEMPERATURE) //
+								.bit(5, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_DISCHARGE_OVER_CURRENT) //
+								.bit(4, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_SYSTEM_LOW_VOLTAGE) //
+								.bit(3, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_CELL_LOW_VOLTAGE) //
+								.bit(2, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_CHARGE_OVER_CURRENT) //
+								.bit(1, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_SYSTEM_OVER_VOLTAGE) //
+								.bit(0, SingleRackVersionB.ChannelId.ALARM_FLAG_STATUS_CELL_OVER_VOLTAGE) //
 						), //
 						m(new BitsWordElement(0x2008, this) //
-								.bit(15, SingleRackChannelId.PROTECT_FLAG_STATUS_DISCHARGE_TEMPERATURE_LOW) //
-								.bit(14, SingleRackChannelId.PROTECT_FLAG_STATUS_DISCHARGE_TEMPERATURE_HIGH) //
-								.bit(13, SingleRackChannelId.PROTECT_FLAG_STATUS_VOLTAGE_DIFFERENCE) //
-								.bit(12, SingleRackChannelId.PROTECT_FLAG_STATUS_INSULATION_LOW) //
-								.bit(11, SingleRackChannelId.PROTECT_FLAG_STATUS_CELL_VOLTAGE_DIFFERENCE) //
-								.bit(10, SingleRackChannelId.PROTECT_FLAG_STATUS_ELECTRODE_TEMPERATURE_HIGH) //
-								.bit(9, SingleRackChannelId.PROTECT_FLAG_STATUS_TEMPERATURE_DIFFERENCE) //
-								.bit(8, SingleRackChannelId.PROTECT_FLAG_STATUS_SOC_LOW) //
-								.bit(7, SingleRackChannelId.PROTECT_FLAG_STATUS_CELL_OVER_TEMPERATURE) //
-								.bit(6, SingleRackChannelId.PROTECT_FLAG_STATUS_CELL_LOW_TEMPERATURE) //
-								.bit(5, SingleRackChannelId.PROTECT_FLAG_STATUS_DISCHARGE_OVER_CURRENT) //
-								.bit(4, SingleRackChannelId.PROTECT_FLAG_STATUS_SYSTEM_LOW_VOLTAGE) //
-								.bit(3, SingleRackChannelId.PROTECT_FLAG_STATUS_CELL_LOW_VOLTAGE) //
-								.bit(2, SingleRackChannelId.PROTECT_FLAG_STATUS_CHARGE_OVER_CURRENT) //
-								.bit(1, SingleRackChannelId.PROTECT_FLAG_STATUS_SYSTEM_OVER_VOLTAGE) //
-								.bit(0, SingleRackChannelId.PROTECT_FLAG_STATUS_CELL_OVER_VOLTAGE) //
+								.bit(15, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_DISCHARGE_TEMPERATURE_LOW) //
+								.bit(14, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_DISCHARGE_TEMPERATURE_HIGH) //
+								.bit(13, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_VOLTAGE_DIFFERENCE) //
+								.bit(12, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_INSULATION_LOW) //
+								.bit(11, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_CELL_VOLTAGE_DIFFERENCE) //
+								.bit(10, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_ELECTRODE_TEMPERATURE_HIGH) //
+								.bit(9, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_TEMPERATURE_DIFFERENCE) //
+								.bit(8, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_SOC_LOW) //
+								.bit(7, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_CELL_OVER_TEMPERATURE) //
+								.bit(6, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_CELL_LOW_TEMPERATURE) //
+								.bit(5, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_DISCHARGE_OVER_CURRENT) //
+								.bit(4, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_SYSTEM_LOW_VOLTAGE) //
+								.bit(3, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_CELL_LOW_VOLTAGE) //
+								.bit(2, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_CHARGE_OVER_CURRENT) //
+								.bit(1, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_SYSTEM_OVER_VOLTAGE) //
+								.bit(0, SingleRackVersionB.ChannelId.PROTECT_FLAG_STATUS_CELL_OVER_VOLTAGE) //
 						), //
-						m(SingleRackChannelId.ALARM_FLAG_REGISTER_1, new UnsignedWordElement(0x2009)), //
-						m(SingleRackChannelId.ALARM_FLAG_REGISTER_2, new UnsignedWordElement(0x200A)), //
-						m(SingleRackChannelId.PROTECT_FLAG_REGISTER_1, new UnsignedWordElement(0x200B)), //
-						m(SingleRackChannelId.PROTECT_FLAG_REGISTER_2, new UnsignedWordElement(0x200C)), //
-						m(SingleRackChannelId.SHORT_CIRCUIT_FUNCTION, new UnsignedWordElement(0x200D)), //
-						m(SingleRackChannelId.TESTING_IO, new UnsignedWordElement(0x200E)), //
-						m(SingleRackChannelId.SOFT_SHUTDOWN, new UnsignedWordElement(0x200F)), //
-						m(SingleRackChannelId.BMS_CONTACTOR_CONTROL, new UnsignedWordElement(0x2010)), //
-						m(SingleRackChannelId.CURRENT_BOX_SELF_CALIBRATION, new UnsignedWordElement(0x2011)), //
-						m(SingleRackChannelId.PCS_ALARM_RESET, new UnsignedWordElement(0x2012)), //
-						m(SingleRackChannelId.INSULATION_SENSOR_FUNCTION, new UnsignedWordElement(0x2013)), //
-						m(SingleRackChannelId.AUTO_SET_SLAVES_ID, new UnsignedWordElement(0x2014)), //
+						m(SingleRackVersionB.ChannelId.ALARM_FLAG_REGISTER_1, new UnsignedWordElement(0x2009)), //
+						m(SingleRackVersionB.ChannelId.ALARM_FLAG_REGISTER_2, new UnsignedWordElement(0x200A)), //
+						m(SingleRackVersionB.ChannelId.PROTECT_FLAG_REGISTER_1, new UnsignedWordElement(0x200B)), //
+						m(SingleRackVersionB.ChannelId.PROTECT_FLAG_REGISTER_2, new UnsignedWordElement(0x200C)), //
+						m(SingleRackVersionB.ChannelId.SHORT_CIRCUIT_FUNCTION, new UnsignedWordElement(0x200D)), //
+						m(SingleRackVersionB.ChannelId.TESTING_IO, new UnsignedWordElement(0x200E)), //
+						m(SingleRackVersionB.ChannelId.SOFT_SHUTDOWN, new UnsignedWordElement(0x200F)), //
+						m(SingleRackVersionB.ChannelId.BMS_CONTACTOR_CONTROL, new UnsignedWordElement(0x2010)), //
+						m(SingleRackVersionB.ChannelId.CURRENT_BOX_SELF_CALIBRATION, new UnsignedWordElement(0x2011)), //
+						m(SingleRackVersionB.ChannelId.PCS_ALARM_RESET, new UnsignedWordElement(0x2012)), //
+						m(SingleRackVersionB.ChannelId.INSULATION_SENSOR_FUNCTION, new UnsignedWordElement(0x2013)), //
+						m(SingleRackVersionB.ChannelId.AUTO_SET_SLAVES_ID, new UnsignedWordElement(0x2014)), //
 						new DummyRegisterElement(0x2015, 0x2018), //
-						m(SingleRackChannelId.AUTO_SET_SLAVES_TEMPERATURE_ID, new UnsignedWordElement(0x2019)), //
-						m(SingleRackChannelId.TRANSPARENT_MASTER, new UnsignedWordElement(0x201A)), //
-						m(SingleRackChannelId.SET_EMS_ADDRESS, new UnsignedWordElement(0x201B)), //
-						m(SingleRackChannelId.EMS_COMMUNICATION_TIMEOUT, new UnsignedWordElement(0x201C)), //
-						m(SingleRackChannelId.SLEEP, new UnsignedWordElement(0x201D)), //
-						m(SingleRackChannelId.VOLTAGE_LOW_PROTECTION, new UnsignedWordElement(0x201E)) //
+						m(SingleRackVersionB.ChannelId.AUTO_SET_SLAVES_TEMPERATURE_ID, new UnsignedWordElement(0x2019)), //
+						m(SingleRackVersionB.ChannelId.TRANSPARENT_MASTER, new UnsignedWordElement(0x201A)), //
+						m(SingleRackVersionB.ChannelId.SET_EMS_ADDRESS, new UnsignedWordElement(0x201B)), //
+						m(SingleRackVersionB.ChannelId.EMS_COMMUNICATION_TIMEOUT, new UnsignedWordElement(0x201C)), //
+						m(SingleRackVersionB.ChannelId.SLEEP, new UnsignedWordElement(0x201D)), //
+						m(SingleRackVersionB.ChannelId.VOLTAGE_LOW_PROTECTION, new UnsignedWordElement(0x201E)) //
 				), //
 
 				// Voltage ranges
 				new FC3ReadRegistersTask(0x2082, Priority.LOW, //						
 						m(new UnsignedWordElement(0x2082)) //
-						.m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_ALARM, ElementToChannelConverter.SCALE_FACTOR_2) //
+						.m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_ALARM, ElementToChannelConverter.SCALE_FACTOR_2) //
 						.m(Battery.ChannelId.CHARGE_MAX_VOLTAGE, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build(), //						
 						new DummyRegisterElement(0x2083, 0x2087),						
 						m(new UnsignedWordElement(0x2088)) //
-						.m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_ALARM, ElementToChannelConverter.SCALE_FACTOR_2) //
+						.m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_ALARM, ElementToChannelConverter.SCALE_FACTOR_2) //
 						.m(Battery.ChannelId.DISCHARGE_MIN_VOLTAGE, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build() //						
 				),
@@ -435,213 +358,213 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 				// Summary state
 				new FC3ReadRegistersTask(0x2100, Priority.LOW,						
 						m(new UnsignedWordElement(0x2100)) //
-						.m(SingleRackChannelId.CLUSTER_1_VOLTAGE, ElementToChannelConverter.SCALE_FACTOR_2) //
+						.m(SingleRackVersionB.ChannelId.CLUSTER_1_VOLTAGE, ElementToChannelConverter.SCALE_FACTOR_2) //
 						.m(Battery.ChannelId.VOLTAGE, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build(), //
 						m(new SignedWordElement(0x2101)) //
-						.m(SingleRackChannelId.CLUSTER_1_CURRENT, ElementToChannelConverter.SCALE_FACTOR_2) //
+						.m(SingleRackVersionB.ChannelId.CLUSTER_1_CURRENT, ElementToChannelConverter.SCALE_FACTOR_2) //
 						.m(Battery.ChannelId.CURRENT, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build(), //
-						m(SingleRackChannelId.CHARGE_INDICATION, new UnsignedWordElement(0x2102)),
+						m(SingleRackVersionB.ChannelId.CHARGE_INDICATION, new UnsignedWordElement(0x2102)),
 						m(Battery.ChannelId.SOC, new UnsignedWordElement(0x2103)),
-						m(SingleRackChannelId.CLUSTER_1_SOH, new UnsignedWordElement(0x2104)),
-						m(SingleRackChannelId.CLUSTER_1_MAX_CELL_VOLTAGE_ID, new UnsignedWordElement(0x2105)), //
+						m(SingleRackVersionB.ChannelId.CLUSTER_1_SOH, new UnsignedWordElement(0x2104)),
+						m(SingleRackVersionB.ChannelId.CLUSTER_1_MAX_CELL_VOLTAGE_ID, new UnsignedWordElement(0x2105)), //
 						m(new SignedWordElement(0x2106)) //
-						.m(SingleRackChannelId.CLUSTER_1_MAX_CELL_VOLTAGE, ElementToChannelConverter.DIRECT_1_TO_1) //
+						.m(SingleRackVersionB.ChannelId.CLUSTER_1_MAX_CELL_VOLTAGE, ElementToChannelConverter.DIRECT_1_TO_1) //
 						.m(Battery.ChannelId.MAX_CELL_VOLTAGE, ElementToChannelConverter.DIRECT_1_TO_1) //
 						.build(), //
-						m(SingleRackChannelId.CLUSTER_1_MIN_CELL_VOLTAGE_ID, new UnsignedWordElement(0x2107)), //
+						m(SingleRackVersionB.ChannelId.CLUSTER_1_MIN_CELL_VOLTAGE_ID, new UnsignedWordElement(0x2107)), //
 						m(new UnsignedWordElement(0x2108)) //
-						.m(SingleRackChannelId.CLUSTER_1_MIN_CELL_VOLTAGE, ElementToChannelConverter.DIRECT_1_TO_1) //
+						.m(SingleRackVersionB.ChannelId.CLUSTER_1_MIN_CELL_VOLTAGE, ElementToChannelConverter.DIRECT_1_TO_1) //
 						.m(Battery.ChannelId.MIN_CELL_VOLTAGE, ElementToChannelConverter.DIRECT_1_TO_1) //
 						.build(), //
-						m(SingleRackChannelId.CLUSTER_1_MAX_CELL_TEMPERATURE_ID, new UnsignedWordElement(0x2109)), //
+						m(SingleRackVersionB.ChannelId.CLUSTER_1_MAX_CELL_TEMPERATURE_ID, new UnsignedWordElement(0x2109)), //
 						m(new SignedWordElement(0x210A)) //
-						.m(SingleRackChannelId.CLUSTER_1_MAX_CELL_TEMPERATURE, ElementToChannelConverter.DIRECT_1_TO_1) //
+						.m(SingleRackVersionB.ChannelId.CLUSTER_1_MAX_CELL_TEMPERATURE, ElementToChannelConverter.DIRECT_1_TO_1) //
 						.m(Battery.ChannelId.MAX_CELL_TEMPERATURE, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build(), //
-						m(SingleRackChannelId.CLUSTER_1_MIN_CELL_TEMPERATURE_ID, new UnsignedWordElement(0x210B)), //
+						m(SingleRackVersionB.ChannelId.CLUSTER_1_MIN_CELL_TEMPERATURE_ID, new UnsignedWordElement(0x210B)), //
 						m(new SignedWordElement(0x210C)) //
-						.m(SingleRackChannelId.CLUSTER_1_MIN_CELL_TEMPERATURE, ElementToChannelConverter.DIRECT_1_TO_1) //
+						.m(SingleRackVersionB.ChannelId.CLUSTER_1_MIN_CELL_TEMPERATURE, ElementToChannelConverter.DIRECT_1_TO_1) //
 						.m(Battery.ChannelId.MIN_CELL_TEMPERATURE, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build(), //
-						m(SingleRackChannelId.MAX_CELL_RESISTANCE_ID, new UnsignedWordElement(0x210D)), //
-						m(SingleRackChannelId.MAX_CELL_RESISTANCE, new UnsignedWordElement(0x210E),
+						m(SingleRackVersionB.ChannelId.MAX_CELL_RESISTANCE_ID, new UnsignedWordElement(0x210D)), //
+						m(SingleRackVersionB.ChannelId.MAX_CELL_RESISTANCE, new UnsignedWordElement(0x210E),
 								ElementToChannelConverter.SCALE_FACTOR_1), //
-						m(SingleRackChannelId.MIN_CELL_RESISTANCE_ID, new UnsignedWordElement(0x210F)), //
-						m(SingleRackChannelId.MIN_CELL_RESISTANCE, new UnsignedWordElement(0x2110),
+						m(SingleRackVersionB.ChannelId.MIN_CELL_RESISTANCE_ID, new UnsignedWordElement(0x210F)), //
+						m(SingleRackVersionB.ChannelId.MIN_CELL_RESISTANCE, new UnsignedWordElement(0x2110),
 								ElementToChannelConverter.SCALE_FACTOR_1), //
-						m(SingleRackChannelId.POSITIVE_INSULATION, new UnsignedWordElement(0x2111)), //
-						m(SingleRackChannelId.NEGATIVE_INSULATION, new UnsignedWordElement(0x2112)), //
-						m(SingleRackChannelId.MAIN_CONTACTOR_FLAG, new UnsignedWordElement(0x2113)), //
+						m(SingleRackVersionB.ChannelId.POSITIVE_INSULATION, new UnsignedWordElement(0x2111)), //
+						m(SingleRackVersionB.ChannelId.NEGATIVE_INSULATION, new UnsignedWordElement(0x2112)), //
+						m(SingleRackVersionB.ChannelId.MAIN_CONTACTOR_FLAG, new UnsignedWordElement(0x2113)), //
 						new DummyRegisterElement(0x2114),
-						m(SingleRackChannelId.ENVIRONMENT_TEMPERATURE, new UnsignedWordElement(0x2115)), //
-						m(SingleRackChannelId.SYSTEM_INSULATION, new UnsignedWordElement(0x2116)), //
-						m(SingleRackChannelId.CELL_VOLTAGE_DIFFERENCE, new UnsignedWordElement(0x2117)), //
-						m(SingleRackChannelId.TOTAL_VOLTAGE_DIFFERENCE, new UnsignedWordElement(0x2118),
+						m(SingleRackVersionB.ChannelId.ENVIRONMENT_TEMPERATURE, new UnsignedWordElement(0x2115)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_INSULATION, new UnsignedWordElement(0x2116)), //
+						m(SingleRackVersionB.ChannelId.CELL_VOLTAGE_DIFFERENCE, new UnsignedWordElement(0x2117)), //
+						m(SingleRackVersionB.ChannelId.TOTAL_VOLTAGE_DIFFERENCE, new UnsignedWordElement(0x2118),
 								ElementToChannelConverter.SCALE_FACTOR_2), //
-						m(SingleRackChannelId.POWER_TEMPERATURE, new UnsignedWordElement(0x2119)), //
-						m(SingleRackChannelId.POWER_SUPPLY_VOLTAGE, new UnsignedWordElement(0x211A)) //
+						m(SingleRackVersionB.ChannelId.POWER_TEMPERATURE, new UnsignedWordElement(0x2119)), //
+						m(SingleRackVersionB.ChannelId.POWER_SUPPLY_VOLTAGE, new UnsignedWordElement(0x211A)) //
 				),
 
 				// Critical state
 				new FC3ReadRegistersTask(0x2140, Priority.HIGH, //
 						m(new BitsWordElement(0x2140, this) //
-								.bit(0, SingleRackChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_HIGH) //
-								.bit(1, SingleRackChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_HIGH) //
-								.bit(2, SingleRackChannelId.ALARM_LEVEL_2_CHA_CURRENT_HIGH) //
-								.bit(3, SingleRackChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_LOW) //
-								.bit(4, SingleRackChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_LOW) //
-								.bit(5, SingleRackChannelId.ALARM_LEVEL_2_DISCHA_CURRENT_HIGH) //
-								.bit(6, SingleRackChannelId.ALARM_LEVEL_2_CELL_CHA_TEMP_HIGH) //
-								.bit(7, SingleRackChannelId.ALARM_LEVEL_2_CELL_CHA_TEMP_LOW) //
-								.bit(8, SingleRackChannelId.ALARM_LEVEL_2_SOC_LOW) //
-								.bit(9, SingleRackChannelId.ALARM_LEVEL_2_TEMPERATURE_DIFFERENCE_HIGH) //
-								.bit(10, SingleRackChannelId.ALARM_LEVEL_2_POLES_TEMPERATURE_DIFFERENCE_HIGH) //
-								.bit(11, SingleRackChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_DIFFERENCE_HIGH) //
-								.bit(12, SingleRackChannelId.ALARM_LEVEL_2_INSULATION_LOW) //
-								.bit(13, SingleRackChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_DIFFERENCE_HIGH) //
-								.bit(14, SingleRackChannelId.ALARM_LEVEL_2_CELL_DISCHA_TEMP_HIGH) //
-								.bit(15, SingleRackChannelId.ALARM_LEVEL_2_CELL_DISCHA_TEMP_LOW) //
+								.bit(0, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_HIGH) //
+								.bit(1, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_HIGH) //
+								.bit(2, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CHA_CURRENT_HIGH) //
+								.bit(3, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_LOW) //
+								.bit(4, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_LOW) //
+								.bit(5, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_DISCHA_CURRENT_HIGH) //
+								.bit(6, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_CHA_TEMP_HIGH) //
+								.bit(7, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_CHA_TEMP_LOW) //
+								.bit(8, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_SOC_LOW) //
+								.bit(9, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_TEMPERATURE_DIFFERENCE_HIGH) //
+								.bit(10, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_POLES_TEMPERATURE_DIFFERENCE_HIGH) //
+								.bit(11, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_VOLTAGE_DIFFERENCE_HIGH) //
+								.bit(12, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_INSULATION_LOW) //
+								.bit(13, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_TOTAL_VOLTAGE_DIFFERENCE_HIGH) //
+								.bit(14, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_DISCHA_TEMP_HIGH) //
+								.bit(15, SingleRackVersionB.ChannelId.ALARM_LEVEL_2_CELL_DISCHA_TEMP_LOW) //
 						), //
 						m(new BitsWordElement(0x2141, this) //
-								.bit(0, SingleRackChannelId.ALARM_LEVEL_1_CELL_VOLTAGE_HIGH) //
-								.bit(1, SingleRackChannelId.ALARM_LEVEL_1_TOTAL_VOLTAGE_HIGH) //
-								.bit(2, SingleRackChannelId.ALARM_LEVEL_1_CHA_CURRENT_HIGH) //
-								.bit(3, SingleRackChannelId.ALARM_LEVEL_1_CELL_VOLTAGE_LOW) //
-								.bit(4, SingleRackChannelId.ALARM_LEVEL_1_TOTAL_VOLTAGE_LOW) //
-								.bit(5, SingleRackChannelId.ALARM_LEVEL_1_DISCHA_CURRENT_HIGH) //
-								.bit(6, SingleRackChannelId.ALARM_LEVEL_1_CELL_CHA_TEMP_HIGH) //
-								.bit(7, SingleRackChannelId.ALARM_LEVEL_1_CELL_CHA_TEMP_LOW) //
-								.bit(8, SingleRackChannelId.ALARM_LEVEL_1_SOC_LOW) //
-								.bit(9, SingleRackChannelId.ALARM_LEVEL_1_CELL_TEMP_DIFF_HIGH) //
-								.bit(10, SingleRackChannelId.ALARM_LEVEL_1_POLE_TEMPERATURE_TOO_HIGH) //
-								.bit(11, SingleRackChannelId.ALARM_LEVEL_1_CELL_VOLTAGE_DIFF_HIGH) //
-								.bit(12, SingleRackChannelId.ALARM_LEVEL_1_INSULATION_LOW) //
-								.bit(13, SingleRackChannelId.ALARM_LEVEL_1_TOTAL_VOLTAGE_DIFF_HIGH) //
-								.bit(14, SingleRackChannelId.ALARM_LEVEL_1_CELL_DISCHA_TEMP_HIGH) //
-								.bit(15, SingleRackChannelId.ALARM_LEVEL_1_CELL_DISCHA_TEMP_LOW) //
+								.bit(0, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_VOLTAGE_HIGH) //
+								.bit(1, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_TOTAL_VOLTAGE_HIGH) //
+								.bit(2, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CHA_CURRENT_HIGH) //
+								.bit(3, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_VOLTAGE_LOW) //
+								.bit(4, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_TOTAL_VOLTAGE_LOW) //
+								.bit(5, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_DISCHA_CURRENT_HIGH) //
+								.bit(6, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_CHA_TEMP_HIGH) //
+								.bit(7, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_CHA_TEMP_LOW) //
+								.bit(8, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_SOC_LOW) //
+								.bit(9, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_TEMP_DIFF_HIGH) //
+								.bit(10, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_POLE_TEMPERATURE_TOO_HIGH) //
+								.bit(11, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_VOLTAGE_DIFF_HIGH) //
+								.bit(12, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_INSULATION_LOW) //
+								.bit(13, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_TOTAL_VOLTAGE_DIFF_HIGH) //
+								.bit(14, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_DISCHA_TEMP_HIGH) //
+								.bit(15, SingleRackVersionB.ChannelId.ALARM_LEVEL_1_CELL_DISCHA_TEMP_LOW) //
 						), //
-						m(SingleRackChannelId.CLUSTER_RUN_STATE, new UnsignedWordElement(0x2142)), //
+						m(SingleRackVersionB.ChannelId.CLUSTER_RUN_STATE, new UnsignedWordElement(0x2142)), //
 
-						m(SingleRackChannelId.MAXIMUM_CELL_VOLTAGE_NUMBER_WHEN_ALARM, new UnsignedWordElement(0x2143)), //
-						m(SingleRackChannelId.MAXIMUM_CELL_VOLTAGE_WHEN_ALARM, new UnsignedWordElement(0x2144)), //
-						m(SingleRackChannelId.MAXIMUM_CELL_VOLTAGE_NUMBER_WHEN_STOPPED,
+						m(SingleRackVersionB.ChannelId.MAXIMUM_CELL_VOLTAGE_NUMBER_WHEN_ALARM, new UnsignedWordElement(0x2143)), //
+						m(SingleRackVersionB.ChannelId.MAXIMUM_CELL_VOLTAGE_WHEN_ALARM, new UnsignedWordElement(0x2144)), //
+						m(SingleRackVersionB.ChannelId.MAXIMUM_CELL_VOLTAGE_NUMBER_WHEN_STOPPED,
 								new UnsignedWordElement(0x2145)), //
-						m(SingleRackChannelId.MAXIMUM_CELL_VOLTAGE_WHEN_STOPPED, new UnsignedWordElement(0x2146)), //
-						m(SingleRackChannelId.MINIMUM_CELL_VOLTAGE_NUMBER_WHEN_ALARM, new UnsignedWordElement(0x2147)), //
-						m(SingleRackChannelId.MINIMUM_CELL_VOLTAGE_WHEN_ALARM, new UnsignedWordElement(0x2148)), //
-						m(SingleRackChannelId.MINIMUM_CELL_VOLTAGE_NUMBER_WHEN_STOPPED,
+						m(SingleRackVersionB.ChannelId.MAXIMUM_CELL_VOLTAGE_WHEN_STOPPED, new UnsignedWordElement(0x2146)), //
+						m(SingleRackVersionB.ChannelId.MINIMUM_CELL_VOLTAGE_NUMBER_WHEN_ALARM, new UnsignedWordElement(0x2147)), //
+						m(SingleRackVersionB.ChannelId.MINIMUM_CELL_VOLTAGE_WHEN_ALARM, new UnsignedWordElement(0x2148)), //
+						m(SingleRackVersionB.ChannelId.MINIMUM_CELL_VOLTAGE_NUMBER_WHEN_STOPPED,
 								new UnsignedWordElement(0x2149)), //
-						m(SingleRackChannelId.MINIMUM_CELL_VOLTAGE_WHEN_STOPPED, new UnsignedWordElement(0x214A)), //
-						m(SingleRackChannelId.OVER_VOLTAGE_VALUE_WHEN_ALARM, new UnsignedWordElement(0x214B)), //
-						m(SingleRackChannelId.OVER_VOLTAGE_VALUE_WHEN_STOPPED, new UnsignedWordElement(0x214C)), //
-						m(SingleRackChannelId.UNDER_VOLTAGE_VALUE_WHEN_ALARM, new UnsignedWordElement(0x214D)), //
-						m(SingleRackChannelId.UNDER_VOLTAGE_VALUE_WHEN_STOPPED, new UnsignedWordElement(0x214E)), //
-						m(SingleRackChannelId.OVER_CHARGE_CURRENT_WHEN_ALARM, new UnsignedWordElement(0x214F)), //
-						m(SingleRackChannelId.OVER_CHARGE_CURRENT_WHEN_STOPPED, new UnsignedWordElement(0x2150)), //
-						m(SingleRackChannelId.OVER_DISCHARGE_CURRENT_WHEN_ALARM, new UnsignedWordElement(0x2151)), //
-						m(SingleRackChannelId.OVER_DISCHARGE_CURRENT_WHEN_STOPPED, new UnsignedWordElement(0x2152)), //
-						m(SingleRackChannelId.NUMBER_OF_TEMPERATURE_WHEN_ALARM, new UnsignedWordElement(0x2153)), //
+						m(SingleRackVersionB.ChannelId.MINIMUM_CELL_VOLTAGE_WHEN_STOPPED, new UnsignedWordElement(0x214A)), //
+						m(SingleRackVersionB.ChannelId.OVER_VOLTAGE_VALUE_WHEN_ALARM, new UnsignedWordElement(0x214B)), //
+						m(SingleRackVersionB.ChannelId.OVER_VOLTAGE_VALUE_WHEN_STOPPED, new UnsignedWordElement(0x214C)), //
+						m(SingleRackVersionB.ChannelId.UNDER_VOLTAGE_VALUE_WHEN_ALARM, new UnsignedWordElement(0x214D)), //
+						m(SingleRackVersionB.ChannelId.UNDER_VOLTAGE_VALUE_WHEN_STOPPED, new UnsignedWordElement(0x214E)), //
+						m(SingleRackVersionB.ChannelId.OVER_CHARGE_CURRENT_WHEN_ALARM, new UnsignedWordElement(0x214F)), //
+						m(SingleRackVersionB.ChannelId.OVER_CHARGE_CURRENT_WHEN_STOPPED, new UnsignedWordElement(0x2150)), //
+						m(SingleRackVersionB.ChannelId.OVER_DISCHARGE_CURRENT_WHEN_ALARM, new UnsignedWordElement(0x2151)), //
+						m(SingleRackVersionB.ChannelId.OVER_DISCHARGE_CURRENT_WHEN_STOPPED, new UnsignedWordElement(0x2152)), //
+						m(SingleRackVersionB.ChannelId.NUMBER_OF_TEMPERATURE_WHEN_ALARM, new UnsignedWordElement(0x2153)), //
 						new DummyRegisterElement(0x2154, 0x215A), //
-						m(SingleRackChannelId.OTHER_ALARM_EQUIPMENT_FAILURE, new UnsignedWordElement(0x215B)), //
+						m(SingleRackVersionB.ChannelId.OTHER_ALARM_EQUIPMENT_FAILURE, new UnsignedWordElement(0x215B)), //
 						new DummyRegisterElement(0x215C, 0x215F), //
 			
 						m(new UnsignedWordElement(0x2160)) //
-						.m(SingleRackChannelId.SYSTEM_MAX_CHARGE_CURRENT, ElementToChannelConverter.SCALE_FACTOR_2) //
+						.m(SingleRackVersionB.ChannelId.SYSTEM_MAX_CHARGE_CURRENT, ElementToChannelConverter.SCALE_FACTOR_2) //
 						.m(Battery.ChannelId.CHARGE_MAX_CURRENT, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build(),  //
 						m(new UnsignedWordElement(0x2161)) //
-						.m(SingleRackChannelId.SYSTEM_MAX_DISCHARGE_CURRENT, ElementToChannelConverter.SCALE_FACTOR_2) //
+						.m(SingleRackVersionB.ChannelId.SYSTEM_MAX_DISCHARGE_CURRENT, ElementToChannelConverter.SCALE_FACTOR_2) //
 						.m(Battery.ChannelId.DISCHARGE_MAX_CURRENT, ElementToChannelConverter.SCALE_FACTOR_MINUS_1) //
 						.build()  //						
 				), //
 
 				// Cluster info
 				new FC3ReadRegistersTask(0x2180, Priority.LOW, //
-						m(SingleRackChannelId.CYCLE_TIME, new UnsignedWordElement(0x2180)), //
-						m(SingleRackChannelId.TOTAL_CAPACITY_HIGH_BITS, new UnsignedWordElement(0x2181)), //
-						m(SingleRackChannelId.TOTAL_CAPACITY_LOW_BITS, new UnsignedWordElement(0x2182)), //
+						m(SingleRackVersionB.ChannelId.CYCLE_TIME, new UnsignedWordElement(0x2180)), //
+						m(SingleRackVersionB.ChannelId.TOTAL_CAPACITY_HIGH_BITS, new UnsignedWordElement(0x2181)), //
+						m(SingleRackVersionB.ChannelId.TOTAL_CAPACITY_LOW_BITS, new UnsignedWordElement(0x2182)), //
 						m(new BitsWordElement(0x2183, this) //
-								.bit(3, SingleRackChannelId.SLAVE_20_COMMUNICATION_ERROR)//
-								.bit(2, SingleRackChannelId.SLAVE_19_COMMUNICATION_ERROR)//
-								.bit(1, SingleRackChannelId.SLAVE_18_COMMUNICATION_ERROR)//
-								.bit(0, SingleRackChannelId.SLAVE_17_COMMUNICATION_ERROR)//
+								.bit(3, SingleRackVersionB.ChannelId.SLAVE_20_COMMUNICATION_ERROR)//
+								.bit(2, SingleRackVersionB.ChannelId.SLAVE_19_COMMUNICATION_ERROR)//
+								.bit(1, SingleRackVersionB.ChannelId.SLAVE_18_COMMUNICATION_ERROR)//
+								.bit(0, SingleRackVersionB.ChannelId.SLAVE_17_COMMUNICATION_ERROR)//
 						), //
 						m(new BitsWordElement(0x2184, this) //
-								.bit(15, SingleRackChannelId.SLAVE_16_COMMUNICATION_ERROR)//
-								.bit(14, SingleRackChannelId.SLAVE_15_COMMUNICATION_ERROR)//
-								.bit(13, SingleRackChannelId.SLAVE_14_COMMUNICATION_ERROR)//
-								.bit(12, SingleRackChannelId.SLAVE_13_COMMUNICATION_ERROR)//
-								.bit(11, SingleRackChannelId.SLAVE_12_COMMUNICATION_ERROR)//
-								.bit(10, SingleRackChannelId.SLAVE_11_COMMUNICATION_ERROR)//
-								.bit(9, SingleRackChannelId.SLAVE_10_COMMUNICATION_ERROR)//
-								.bit(8, SingleRackChannelId.SLAVE_9_COMMUNICATION_ERROR)//
-								.bit(7, SingleRackChannelId.SLAVE_8_COMMUNICATION_ERROR)//
-								.bit(6, SingleRackChannelId.SLAVE_7_COMMUNICATION_ERROR)//
-								.bit(5, SingleRackChannelId.SLAVE_6_COMMUNICATION_ERROR)//
-								.bit(4, SingleRackChannelId.SLAVE_5_COMMUNICATION_ERROR)//
-								.bit(3, SingleRackChannelId.SLAVE_4_COMMUNICATION_ERROR)//
-								.bit(2, SingleRackChannelId.SLAVE_3_COMMUNICATION_ERROR)//
-								.bit(1, SingleRackChannelId.SLAVE_2_COMMUNICATION_ERROR)//
-								.bit(0, SingleRackChannelId.SLAVE_1_COMMUNICATION_ERROR)//
+								.bit(15, SingleRackVersionB.ChannelId.SLAVE_16_COMMUNICATION_ERROR)//
+								.bit(14, SingleRackVersionB.ChannelId.SLAVE_15_COMMUNICATION_ERROR)//
+								.bit(13, SingleRackVersionB.ChannelId.SLAVE_14_COMMUNICATION_ERROR)//
+								.bit(12, SingleRackVersionB.ChannelId.SLAVE_13_COMMUNICATION_ERROR)//
+								.bit(11, SingleRackVersionB.ChannelId.SLAVE_12_COMMUNICATION_ERROR)//
+								.bit(10, SingleRackVersionB.ChannelId.SLAVE_11_COMMUNICATION_ERROR)//
+								.bit(9, SingleRackVersionB.ChannelId.SLAVE_10_COMMUNICATION_ERROR)//
+								.bit(8, SingleRackVersionB.ChannelId.SLAVE_9_COMMUNICATION_ERROR)//
+								.bit(7, SingleRackVersionB.ChannelId.SLAVE_8_COMMUNICATION_ERROR)//
+								.bit(6, SingleRackVersionB.ChannelId.SLAVE_7_COMMUNICATION_ERROR)//
+								.bit(5, SingleRackVersionB.ChannelId.SLAVE_6_COMMUNICATION_ERROR)//
+								.bit(4, SingleRackVersionB.ChannelId.SLAVE_5_COMMUNICATION_ERROR)//
+								.bit(3, SingleRackVersionB.ChannelId.SLAVE_4_COMMUNICATION_ERROR)//
+								.bit(2, SingleRackVersionB.ChannelId.SLAVE_3_COMMUNICATION_ERROR)//
+								.bit(1, SingleRackVersionB.ChannelId.SLAVE_2_COMMUNICATION_ERROR)//
+								.bit(0, SingleRackVersionB.ChannelId.SLAVE_1_COMMUNICATION_ERROR)//
 						), //
 						m(new BitsWordElement(0x2185, this) //
-								.bit(0, SingleRackChannelId.FAILURE_SAMPLING_WIRE)//
-								.bit(1, SingleRackChannelId.FAILURE_CONNECTOR_WIRE)//
-								.bit(2, SingleRackChannelId.FAILURE_LTC6803)//
-								.bit(3, SingleRackChannelId.FAILURE_VOLTAGE_SAMPLING)//
-								.bit(4, SingleRackChannelId.FAILURE_TEMP_SAMPLING)//
-								.bit(5, SingleRackChannelId.FAILURE_TEMP_SENSOR)//
-								.bit(6, SingleRackChannelId.FAILURE_GR_T)//
-								.bit(7, SingleRackChannelId.FAILURE_PCB)//
-								.bit(8, SingleRackChannelId.FAILURE_BALANCING_MODULE)//
-								.bit(9, SingleRackChannelId.FAILURE_TEMP_SAMPLING_LINE)//
-								.bit(10, SingleRackChannelId.FAILURE_INTRANET_COMMUNICATION)//
-								.bit(11, SingleRackChannelId.FAILURE_EEPROM)//
-								.bit(12, SingleRackChannelId.FAILURE_INITIALIZATION)//
+								.bit(0, SingleRackVersionB.ChannelId.FAILURE_SAMPLING_WIRE)//
+								.bit(1, SingleRackVersionB.ChannelId.FAILURE_CONNECTOR_WIRE)//
+								.bit(2, SingleRackVersionB.ChannelId.FAILURE_LTC6803)//
+								.bit(3, SingleRackVersionB.ChannelId.FAILURE_VOLTAGE_SAMPLING)//
+								.bit(4, SingleRackVersionB.ChannelId.FAILURE_TEMP_SAMPLING)//
+								.bit(5, SingleRackVersionB.ChannelId.FAILURE_TEMP_SENSOR)//
+								.bit(6, SingleRackVersionB.ChannelId.FAILURE_GR_T)//
+								.bit(7, SingleRackVersionB.ChannelId.FAILURE_PCB)//
+								.bit(8, SingleRackVersionB.ChannelId.FAILURE_BALANCING_MODULE)//
+								.bit(9, SingleRackVersionB.ChannelId.FAILURE_TEMP_SAMPLING_LINE)//
+								.bit(10, SingleRackVersionB.ChannelId.FAILURE_INTRANET_COMMUNICATION)//
+								.bit(11, SingleRackVersionB.ChannelId.FAILURE_EEPROM)//
+								.bit(12, SingleRackVersionB.ChannelId.FAILURE_INITIALIZATION)//
 						), //
-						m(SingleRackChannelId.SYSTEM_TIME_HIGH, new UnsignedWordElement(0x2186)), //
-						m(SingleRackChannelId.SYSTEM_TIME_LOW, new UnsignedWordElement(0x2187)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_TIME_HIGH, new UnsignedWordElement(0x2186)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_TIME_LOW, new UnsignedWordElement(0x2187)), //
 						new DummyRegisterElement(0x2188, 0x218E), //
-						m(SingleRackChannelId.LAST_TIME_CHARGE_CAPACITY_LOW_BITS, new UnsignedWordElement(0x218F),
+						m(SingleRackVersionB.ChannelId.LAST_TIME_CHARGE_CAPACITY_LOW_BITS, new UnsignedWordElement(0x218F),
 								ElementToChannelConverter.SCALE_FACTOR_1), //
-						m(SingleRackChannelId.LAST_TIME_CHARGE_END_TIME_HIGH_BITS, new UnsignedWordElement(0x2190)), //
-						m(SingleRackChannelId.LAST_TIME_CHARGE_END_TIME_LOW_BITS, new UnsignedWordElement(0x2191)), //
+						m(SingleRackVersionB.ChannelId.LAST_TIME_CHARGE_END_TIME_HIGH_BITS, new UnsignedWordElement(0x2190)), //
+						m(SingleRackVersionB.ChannelId.LAST_TIME_CHARGE_END_TIME_LOW_BITS, new UnsignedWordElement(0x2191)), //
 						new DummyRegisterElement(0x2192), //
-						m(SingleRackChannelId.LAST_TIME_DISCHARGE_CAPACITY_LOW_BITS, new UnsignedWordElement(0x2193),
+						m(SingleRackVersionB.ChannelId.LAST_TIME_DISCHARGE_CAPACITY_LOW_BITS, new UnsignedWordElement(0x2193),
 								ElementToChannelConverter.SCALE_FACTOR_1), //
-						m(SingleRackChannelId.LAST_TIME_DISCHARGE_END_TIME_HIGH_BITS, new UnsignedWordElement(0x2194)), //
-						m(SingleRackChannelId.LAST_TIME_DISCHARGE_END_TIME_LOW_BITS, new UnsignedWordElement(0x2195)), //
-						m(SingleRackChannelId.CELL_OVER_VOLTAGE_STOP_TIMES, new UnsignedWordElement(0x2196)), //
-						m(SingleRackChannelId.BATTERY_OVER_VOLTAGE_STOP_TIMES, new UnsignedWordElement(0x2197)), //
-						m(SingleRackChannelId.BATTERY_CHARGE_OVER_CURRENT_STOP_TIMES, new UnsignedWordElement(0x2198)), //
-						m(SingleRackChannelId.CELL_VOLTAGE_LOW_STOP_TIMES, new UnsignedWordElement(0x2199)), //
-						m(SingleRackChannelId.BATTERY_VOLTAGE_LOW_STOP_TIMES, new UnsignedWordElement(0x219A)), //
-						m(SingleRackChannelId.BATTERY_DISCHARGE_OVER_CURRENT_STOP_TIMES,
+						m(SingleRackVersionB.ChannelId.LAST_TIME_DISCHARGE_END_TIME_HIGH_BITS, new UnsignedWordElement(0x2194)), //
+						m(SingleRackVersionB.ChannelId.LAST_TIME_DISCHARGE_END_TIME_LOW_BITS, new UnsignedWordElement(0x2195)), //
+						m(SingleRackVersionB.ChannelId.CELL_OVER_VOLTAGE_STOP_TIMES, new UnsignedWordElement(0x2196)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_OVER_VOLTAGE_STOP_TIMES, new UnsignedWordElement(0x2197)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_CHARGE_OVER_CURRENT_STOP_TIMES, new UnsignedWordElement(0x2198)), //
+						m(SingleRackVersionB.ChannelId.CELL_VOLTAGE_LOW_STOP_TIMES, new UnsignedWordElement(0x2199)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_VOLTAGE_LOW_STOP_TIMES, new UnsignedWordElement(0x219A)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_DISCHARGE_OVER_CURRENT_STOP_TIMES,
 								new UnsignedWordElement(0x219B)), //
-						m(SingleRackChannelId.BATTERY_OVER_TEMPERATURE_STOP_TIMES, new UnsignedWordElement(0x219C)), //
-						m(SingleRackChannelId.BATTERY_TEMPERATURE_LOW_STOP_TIMES, new UnsignedWordElement(0x219D)), //
-						m(SingleRackChannelId.CELL_OVER_VOLTAGE_ALARM_TIMES, new UnsignedWordElement(0x219E)), //
-						m(SingleRackChannelId.BATTERY_OVER_VOLTAGE_ALARM_TIMES, new UnsignedWordElement(0x219F)), //
-						m(SingleRackChannelId.BATTERY_CHARGE_OVER_CURRENT_ALARM_TIMES, new UnsignedWordElement(0x21A0)), //
-						m(SingleRackChannelId.CELL_VOLTAGE_LOW_ALARM_TIMES, new UnsignedWordElement(0x21A1)), //
-						m(SingleRackChannelId.BATTERY_VOLTAGE_LOW_ALARM_TIMES, new UnsignedWordElement(0x21A2)), //
-						m(SingleRackChannelId.BATTERY_DISCHARGE_OVER_CURRENT_ALARM_TIMES,
+						m(SingleRackVersionB.ChannelId.BATTERY_OVER_TEMPERATURE_STOP_TIMES, new UnsignedWordElement(0x219C)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_TEMPERATURE_LOW_STOP_TIMES, new UnsignedWordElement(0x219D)), //
+						m(SingleRackVersionB.ChannelId.CELL_OVER_VOLTAGE_ALARM_TIMES, new UnsignedWordElement(0x219E)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_OVER_VOLTAGE_ALARM_TIMES, new UnsignedWordElement(0x219F)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_CHARGE_OVER_CURRENT_ALARM_TIMES, new UnsignedWordElement(0x21A0)), //
+						m(SingleRackVersionB.ChannelId.CELL_VOLTAGE_LOW_ALARM_TIMES, new UnsignedWordElement(0x21A1)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_VOLTAGE_LOW_ALARM_TIMES, new UnsignedWordElement(0x21A2)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_DISCHARGE_OVER_CURRENT_ALARM_TIMES,
 								new UnsignedWordElement(0x21A3)), //
-						m(SingleRackChannelId.BATTERY_OVER_TEMPERATURE_ALARM_TIMES, new UnsignedWordElement(0x21A4)), //
-						m(SingleRackChannelId.BATTERY_TEMPERATURE_LOW_ALARM_TIMES, new UnsignedWordElement(0x21A5)), //
-						m(SingleRackChannelId.SYSTEM_SHORT_CIRCUIT_PROTECTION_TIMES, new UnsignedWordElement(0x21A6)), //
-						m(SingleRackChannelId.SYSTEM_GR_OVER_TEMPERATURE_STOP_TIMES, new UnsignedWordElement(0x21A7)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_OVER_TEMPERATURE_ALARM_TIMES, new UnsignedWordElement(0x21A4)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_TEMPERATURE_LOW_ALARM_TIMES, new UnsignedWordElement(0x21A5)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_SHORT_CIRCUIT_PROTECTION_TIMES, new UnsignedWordElement(0x21A6)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_GR_OVER_TEMPERATURE_STOP_TIMES, new UnsignedWordElement(0x21A7)), //
 						new DummyRegisterElement(0x21A8), //
-						m(SingleRackChannelId.SYSTEM_GR_OVER_TEMPERATURE_ALARM_TIMES, new UnsignedWordElement(0x21A9)), //
+						m(SingleRackVersionB.ChannelId.SYSTEM_GR_OVER_TEMPERATURE_ALARM_TIMES, new UnsignedWordElement(0x21A9)), //
 						new DummyRegisterElement(0x21AA), //
-						m(SingleRackChannelId.BATTERY_VOLTAGE_DIFFERENCE_ALARM_TIMES, new UnsignedWordElement(0x21AB)), //
-						m(SingleRackChannelId.BATTERY_VOLTAGE_DIFFERENCE_STOP_TIMES, new UnsignedWordElement(0x21AC)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_VOLTAGE_DIFFERENCE_ALARM_TIMES, new UnsignedWordElement(0x21AB)), //
+						m(SingleRackVersionB.ChannelId.BATTERY_VOLTAGE_DIFFERENCE_STOP_TIMES, new UnsignedWordElement(0x21AC)), //
 						new DummyRegisterElement(0x21AD, 0x21B3), //
-						m(SingleRackChannelId.SLAVE_TEMPERATURE_COMMUNICATION_ERROR_HIGH,
+						m(SingleRackVersionB.ChannelId.SLAVE_TEMPERATURE_COMMUNICATION_ERROR_HIGH,
 								new UnsignedWordElement(0x21B4)), //
-						m(SingleRackChannelId.SLAVE_TEMPERATURE_COMMUNICATION_ERROR_LOW,
+						m(SingleRackVersionB.ChannelId.SLAVE_TEMPERATURE_COMMUNICATION_ERROR_LOW,
 								new UnsignedWordElement(0x21B5)) //
 				) //
 		); //
@@ -651,247 +574,247 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 			// Add tasks to read/write work and warn parameters
 			// Stop parameter
 			Task writeStopParameters = new FC16WriteRegistersTask(0x2040, //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_PROTECTION, new UnsignedWordElement(0x2040)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2041)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_PROTECTION, new UnsignedWordElement(0x2040)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2041)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_PROTECTION,
 							new UnsignedWordElement(0x2042), ElementToChannelConverter.SCALE_FACTOR_2),
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2043),
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2043),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_PROTECTION,
 							new UnsignedWordElement(0x2044), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x2045), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_PROTECTION,
 							new UnsignedWordElement(0x2046)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2047)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2047)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_PROTECTION,
 							new UnsignedWordElement(0x2048), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2049),
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2049),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_PROTECTION,
 							new UnsignedWordElement(0x204A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x204B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_PROTECTION,
 							new UnsignedWordElement(0x204C)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x204D)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_PROTECTION,
 							new UnsignedWordElement(0x204E)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x204F)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION, new UnsignedWordElement(0x2050)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION_RECOVER, new UnsignedWordElement(0x2051)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION, new UnsignedWordElement(0x2052)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION_RECOVER, new UnsignedWordElement(0x2053)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION, new UnsignedWordElement(0x2050)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION_RECOVER, new UnsignedWordElement(0x2051)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION, new UnsignedWordElement(0x2052)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION_RECOVER, new UnsignedWordElement(0x2053)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION,
 							new UnsignedWordElement(0x2054)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2055)), //
-					m(SingleRackChannelId.STOP_PARAMETER_INSULATION_PROTECTION, new UnsignedWordElement(0x2056)), //
-					m(SingleRackChannelId.STOP_PARAMETER_INSULATION_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_INSULATION_PROTECTION, new UnsignedWordElement(0x2056)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_INSULATION_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2057)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION,
 							new UnsignedWordElement(0x2058)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2059)), //
-					m(SingleRackChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION,
 							new UnsignedWordElement(0x205A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x205B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION,
 							new UnsignedWordElement(0x205C)), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x205D)), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION,
 							new UnsignedWordElement(0x205E)), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x205F)), //
-					m(SingleRackChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION,
 							new UnsignedWordElement(0x2060)), //
-					m(SingleRackChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2061)) //
 			);
 
 			// Warn parameter
 			Task writeWarnParameters = new FC16WriteRegistersTask(0x2080, //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_ALARM, new UnsignedWordElement(0x2080)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2081)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_ALARM, new UnsignedWordElement(0x2082),
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_ALARM, new UnsignedWordElement(0x2080)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2081)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_ALARM, new UnsignedWordElement(0x2082),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2083),
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2083),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_ALARM,
 							new UnsignedWordElement(0x2084), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x2085), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_ALARM, new UnsignedWordElement(0x2086)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2087)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_ALARM, new UnsignedWordElement(0x2088),
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_ALARM, new UnsignedWordElement(0x2086)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2087)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_ALARM, new UnsignedWordElement(0x2088),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2089),
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2089),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_ALARM,
 							new UnsignedWordElement(0x208A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x208B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208C)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208C)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x208D)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208E)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208E)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x208F)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_LOW_ALARM, new UnsignedWordElement(0x2090)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_LOW_ALARM_RECOVER, new UnsignedWordElement(0x2091)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_HIGH_ALARM, new UnsignedWordElement(0x2092)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_HIGH_ALARM_RECOVER, new UnsignedWordElement(0x2093)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_LOW_ALARM, new UnsignedWordElement(0x2090)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_LOW_ALARM_RECOVER, new UnsignedWordElement(0x2091)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_HIGH_ALARM, new UnsignedWordElement(0x2092)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_HIGH_ALARM_RECOVER, new UnsignedWordElement(0x2093)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM,
 							new UnsignedWordElement(0x2094)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM_RECOVER,
 							new UnsignedWordElement(0x2095)), //
-					m(SingleRackChannelId.WARN_PARAMETER_INSULATION_ALARM, new UnsignedWordElement(0x2096)), //
-					m(SingleRackChannelId.WARN_PARAMETER_INSULATION_ALARM_RECOVER, new UnsignedWordElement(0x2097)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_INSULATION_ALARM, new UnsignedWordElement(0x2096)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_INSULATION_ALARM_RECOVER, new UnsignedWordElement(0x2097)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM,
 							new UnsignedWordElement(0x2098)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
 							new UnsignedWordElement(0x2099)), //
-					m(SingleRackChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM,
 							new UnsignedWordElement(0x209A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
 							new UnsignedWordElement(0x209B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM,
 							new UnsignedWordElement(0x209C)), //
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM_RECOVER,
 							new UnsignedWordElement(0x209D)), //
 					new DummyRegisterElement(0x209E),
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM,
 							new UnsignedWordElement(0x209F)), //
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM_RECOVER,
 							new UnsignedWordElement(0x20A0)), //
-					m(SingleRackChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM, new UnsignedWordElement(0x20A1)), //
-					m(SingleRackChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM, new UnsignedWordElement(0x20A1)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM_RECOVER,
 							new UnsignedWordElement(0x20A2)) //
 			);
 
 			// Stop parameter
 			Task readStopParameters = new FC3ReadRegistersTask(0x2040, Priority.LOW, //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_PROTECTION, new UnsignedWordElement(0x2040)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2041)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_PROTECTION, new UnsignedWordElement(0x2040)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2041)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_PROTECTION,
 							new UnsignedWordElement(0x2042), ElementToChannelConverter.SCALE_FACTOR_2),
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2043),
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2043),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_PROTECTION,
 							new UnsignedWordElement(0x2044), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x2045), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_PROTECTION,
 							new UnsignedWordElement(0x2046)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2047)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2047)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_PROTECTION,
 							new UnsignedWordElement(0x2048), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2049),
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2049),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_PROTECTION,
 							new UnsignedWordElement(0x204A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x204B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_PROTECTION,
 							new UnsignedWordElement(0x204C)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x204D)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_PROTECTION,
 							new UnsignedWordElement(0x204E)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x204F)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION, new UnsignedWordElement(0x2050)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION_RECOVER, new UnsignedWordElement(0x2051)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION, new UnsignedWordElement(0x2052)), //
-					m(SingleRackChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION_RECOVER, new UnsignedWordElement(0x2053)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION, new UnsignedWordElement(0x2050)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_LOW_PROTECTION_RECOVER, new UnsignedWordElement(0x2051)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION, new UnsignedWordElement(0x2052)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_SOC_HIGH_PROTECTION_RECOVER, new UnsignedWordElement(0x2053)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION,
 							new UnsignedWordElement(0x2054)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2055)), //
-					m(SingleRackChannelId.STOP_PARAMETER_INSULATION_PROTECTION, new UnsignedWordElement(0x2056)), //
-					m(SingleRackChannelId.STOP_PARAMETER_INSULATION_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_INSULATION_PROTECTION, new UnsignedWordElement(0x2056)), //
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_INSULATION_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2057)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION,
 							new UnsignedWordElement(0x2058)), //
-					m(SingleRackChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_CELL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2059)), //
-					m(SingleRackChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION,
 							new UnsignedWordElement(0x205A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x205B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION,
 							new UnsignedWordElement(0x205C)), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x205D)), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION,
 							new UnsignedWordElement(0x205E)), //
-					m(SingleRackChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_DISCHARGE_TEMPERATURE_LOW_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x205F)), //
-					m(SingleRackChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION,
 							new UnsignedWordElement(0x2060)), //
-					m(SingleRackChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION_RECOVER,
+					m(SingleRackVersionB.ChannelId.STOP_PARAMETER_TEMPERATURE_DIFFERENCE_PROTECTION_RECOVER,
 							new UnsignedWordElement(0x2061)) //
 			);
 
 			// Warn parameter
 			Task readWarnParameters = new FC3ReadRegistersTask(0x2080, Priority.LOW, //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_ALARM, new UnsignedWordElement(0x2080)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2081)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_ALARM, new UnsignedWordElement(0x2080)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2081)), //
 					new DummyRegisterElement(0x2082),
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2083),
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_OVER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2083),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_ALARM,
 							new UnsignedWordElement(0x2084), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_CHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x2085), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_ALARM, new UnsignedWordElement(0x2086)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2087)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_ALARM, new UnsignedWordElement(0x2086)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2087)), //
 					new DummyRegisterElement(0x2088),
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2089),
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_UNDER_VOLTAGE_RECOVER, new UnsignedWordElement(0x2089),
 							ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_ALARM,
 							new UnsignedWordElement(0x208A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SYSTEM_DISCHARGE_OVER_CURRENT_RECOVER,
 							new UnsignedWordElement(0x208B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208C)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208C)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_OVER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x208D)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208E)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_ALARM, new UnsignedWordElement(0x208E)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_UNDER_TEMPERATURE_RECOVER,
 							new UnsignedWordElement(0x208F)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_LOW_ALARM, new UnsignedWordElement(0x2090)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_LOW_ALARM_RECOVER, new UnsignedWordElement(0x2091)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_HIGH_ALARM, new UnsignedWordElement(0x2092)), //
-					m(SingleRackChannelId.WARN_PARAMETER_SOC_HIGH_ALARM_RECOVER, new UnsignedWordElement(0x2093)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_LOW_ALARM, new UnsignedWordElement(0x2090)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_LOW_ALARM_RECOVER, new UnsignedWordElement(0x2091)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_HIGH_ALARM, new UnsignedWordElement(0x2092)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_SOC_HIGH_ALARM_RECOVER, new UnsignedWordElement(0x2093)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM,
 							new UnsignedWordElement(0x2094)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CONNECTOR_TEMPERATURE_HIGH_ALARM_RECOVER,
 							new UnsignedWordElement(0x2095)), //
-					m(SingleRackChannelId.WARN_PARAMETER_INSULATION_ALARM, new UnsignedWordElement(0x2096)), //
-					m(SingleRackChannelId.WARN_PARAMETER_INSULATION_ALARM_RECOVER, new UnsignedWordElement(0x2097)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_INSULATION_ALARM, new UnsignedWordElement(0x2096)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_INSULATION_ALARM_RECOVER, new UnsignedWordElement(0x2097)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM,
 							new UnsignedWordElement(0x2098)), //
-					m(SingleRackChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_CELL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
 							new UnsignedWordElement(0x2099)), //
-					m(SingleRackChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM,
 							new UnsignedWordElement(0x209A), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TOTAL_VOLTAGE_DIFFERENCE_ALARM_RECOVER,
 							new UnsignedWordElement(0x209B), ElementToChannelConverter.SCALE_FACTOR_2), //
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM,
 							new UnsignedWordElement(0x209C)), //
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_HIGH_ALARM_RECOVER,
 							new UnsignedWordElement(0x209D)), //
 					new DummyRegisterElement(0x209E),
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM,
 							new UnsignedWordElement(0x209F)), //
-					m(SingleRackChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_DISCHARGE_TEMPERATURE_LOW_ALARM_RECOVER,
 							new UnsignedWordElement(0x20A0)), //
-					m(SingleRackChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM, new UnsignedWordElement(0x20A1)), //
-					m(SingleRackChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM_RECOVER,
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM, new UnsignedWordElement(0x20A1)), //
+					m(SingleRackVersionB.ChannelId.WARN_PARAMETER_TEMPERATURE_DIFFERENCE_ALARM_RECOVER,
 							new UnsignedWordElement(0x20A2)) //
 			);
 
@@ -936,42 +859,4 @@ public class SingleRackVersionBImpl extends AbstractOpenemsModbusComponent
 
 		return protocol;
 	}
-
-	@Override
-	public ModbusSlaveTable getModbusSlaveTable(AccessMode accessMode) {
-		return new ModbusSlaveTable(//
-				OpenemsComponent.getModbusSlaveNatureTable(accessMode), //
-				Battery.getModbusSlaveNatureTable(accessMode) //
-		);
-	}
-
-
-	@Override
-	public void setStartStop(StartStop value) {
-		if (this.startStopTarget.getAndSet(value) != value) {
-			// Set only if value changed
-			this.stateMachine.forceNextState(State.UNDEFINED);
-		}
-	}
-
-	@Override
-	public StartStop getStartStopTarget() {
-		switch (this.config.startStop()) {
-		case AUTO:
-			// read StartStop-Channel
-			return this.startStopTarget.get();
-
-		case START:
-			// force START
-			return StartStop.START;
-
-		case STOP:
-			// force STOP
-			return StartStop.STOP;
-		}
-
-		assert false;
-		return StartStop.UNDEFINED; // can never happen
-	}
-
 }
