+++ io.openems.edge.battery.soltaro/src/io/openems/edge/battery/soltaro/single/versionb/statemachine/Context.java
@@ -1,7 +1,7 @@
 package io.openems.edge.battery.soltaro.single.versionb.statemachine;
 
-import io.openems.edge.battery.soltaro.single.versionb.SingleRackVersionB;
 import io.openems.edge.battery.soltaro.single.versionb.Config;
+import io.openems.edge.battery.soltaro.single.versionb.SingleRackVersionB;
 
 public class Context {
 	protected final SingleRackVersionB component;
